% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registers.R, R/registers_package.R
\docType{package}
\name{registers}
\alias{registers}
\alias{show_registers}
\alias{print.registers}
\alias{reset_registers}
\alias{set_registers_option}
\alias{get_registers_option}
\alias{package-registers}
\title{Registers environment}
\usage{
registers()

show_registers()

\method{print}{registers}(x, ...)

reset_registers()

set_registers_option(...)

get_registers_option(option)
}
\arguments{
\item{x}{An object of class \code{registers}}

\item{...}{Additional arguments}

\item{option}{A named registers option to retrieve}
}
\value{
Invisible \code{NULL}
}
\description{
See registers

Registers for R
}
\examples{

registers()
set_register(a <- rnorm(1), "a")
registers()
ls(registers())

# in RStudio, go to Tools > Modify Keyboard Shortcuts...
# set Show registers to Alt-Shift-Cmd-R, Apply, then try it

reset_registers()
registers()

get_registers_option("clobber")
set_registers_option("clobber" = FALSE)
get_registers_option("clobber")
set_registers_option("clobber" = FALSE)
get_registers_option("clobber")
reset_registers()
get_registers_option("clobber")

}
\seealso{
\code{\link[=set_register]{set_register()}}
}
