% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registers_env.R
\name{registers_env}
\alias{registers_env}
\alias{reset_registers_env}
\alias{print.registers_env}
\title{Registers environment}
\usage{
registers_env()

reset_registers_env()

\method{print}{registers_env}(x, ...)
}
\arguments{
\item{x}{An object of class \code{registers_env}}

\item{...}{Additional arguments to pass to \code{\link[=cat]{cat()}}}
}
\value{
Invisible \code{NULL}
}
\description{
See registers
}
\examples{

registers_env()
a <- 1
set_focus("a")
registers_env()

}
