% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_register.R
\name{set_register}
\alias{set_register}
\alias{set_focus}
\alias{set_focus_highlighted}
\alias{set_focus_window}
\alias{glimpse_focus}
\alias{glimpse_highlighted}
\title{Registers}
\usage{
set_focus(name)

set_focus_highlighted()

set_focus_window()

glimpse_focus()

glimpse_highlighted()
}
\arguments{
\item{name}{The binding name to focus on, as a string}
}
\value{
Invisible \code{NULL}
}
\description{
Set registers and register actions
}
\details{
In Vim-speak, \code{focus} is the unnamed register.
}
\examples{

registers()

set_focus("cars")
glimpse_focus()

# in RStudio, go to Tools > Modify Keyboard Shortcuts...
# next to Set focus - highlighted, use Alt-Shift-Cmd-F
# next to Set focus - window, use Alt-F
# next to Glimpse focus, use Shift-Cmd-G

# press Shift-Cmd-G

airquality
# highlight airquality above and push Alt-Shift-Cmd-F
# press Shift-Cmd-G

# in Tools > Modify Keyboard Shortcuts...,
# set Glimpse highlighted to Alt-Shift-H, then highlight the line below and
# press Shift-Cmd-H
subset(cars, speed > 10)

# interesting use case: generate a random subsample (that has a desired property)
# and save it. highlight the whole next line and Shift-Cmd-H
(cars_sample <- cars[sample(nrow(cars), 5),])
cars_sample




}
