% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spy_inline.R
\name{spy_inline}
\alias{spy_inline}
\alias{spy_inline.character}
\alias{spy_inline.numeric}
\alias{spy_inline.logical}
\alias{spy_inline.function}
\alias{spy_inline.environment}
\alias{spy_inline.tbl_df}
\alias{spy_inline.data.frame}
\alias{spy_inline.list}
\alias{spy_inline.default}
\title{Registers inline summaries}
\usage{
spy_inline(x, ...)

\method{spy_inline}{character}(x, ...)

\method{spy_inline}{numeric}(x, ...)

\method{spy_inline}{logical}(x, ...)

\method{spy_inline}{`function`}(x, ...)

\method{spy_inline}{environment}(x, ...)

\method{spy_inline}{tbl_df}(x, ...)

\method{spy_inline}{data.frame}(x, ...)

\method{spy_inline}{list}(x, ...)

\method{spy_inline}{default}(x, ...)
}
\arguments{
\item{x}{An object of class \code{registers}}

\item{...}{Additional arguments to pass to \code{\link[=cat]{cat()}}}
}
\value{
\code{x}
}
\description{
Registers inline summaries
}
\examples{

spy_inline(c("foo", "bar"))
spy_inline(rnorm(3))
spy_inline(1:3)
spy_inline(c(TRUE, FALSE, TRUE))

spy_inline(cars)
spy_inline(list(a = 1, b = 2))
spy_inline(list(1, 2))

spy_inline(function(x, y) x + y)
spy_inline(function(x, y) {
  z <- x^2
  z + y
})

e <- new.env()
assign("a", 1, envir = e)
e
ls(e)
e$a
spy_inline(e)
g <- new.env(parent = e)
spy_inline(g)
spy_inline(asNamespace("dplyr"))




}
