% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\docType{data}
\name{tracking}
\alias{tracking}
\alias{history_symbols}
\alias{track_assignment}
\alias{untrack_assignment}
\alias{pause_tracking}
\alias{resume_tracking}
\title{Track assignment}
\format{
An object of class \code{character} of length 2.
}
\usage{
history_symbols

track_assignment(symb = c("<-", "<+"))

untrack_assignment()

pause_tracking()

resume_tracking()
}
\arguments{
\item{symb}{Symbol to use for tracked assignment, either \code{"<-"} (default) or
\code{"<+"}}
}
\value{
Invisible \code{NULL}
}
\description{
Track assignment
}
\examples{

track_assignment()
ls()
ls(all.names = TRUE)
registers()
a <- "first"
.
b <- "second"
.
..
registers()
c <- "third"
.
..


# composing with history registers
registers()
set_register(spy(), "s")
a <- "first"
b <- "second"
c <- "third"
c(., ..)
ring_register(".s")
ring_register(":s")


if (FALSE) { # oddities

  # assigning to history symbols errors but does update history
  ..
  .. <- 10
  .

  # replacement functions don't work in the usual way but do update history
  x <- 1:3
  .
  x[1] <- 2
  x
  .

}
untrack_assignment()
}
\keyword{datasets}
