% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registers.R, R/registers_package.R
\docType{package}
\name{registers}
\alias{registers}
\alias{show_registers}
\alias{reset_registers}
\alias{print.registers}
\alias{track_assignment}
\alias{untrack_assignment}
\alias{register_inline_summary}
\alias{register_inline_summary.character}
\alias{register_inline_summary.numeric}
\alias{register_inline_summary.logical}
\alias{register_inline_summary.function}
\alias{register_inline_summary.environment}
\alias{register_inline_summary.tbl_df}
\alias{register_inline_summary.data.frame}
\alias{register_inline_summary.default}
\alias{package-registers}
\title{Registers environment}
\usage{
registers()

show_registers()

reset_registers()

\method{print}{registers}(x, ...)

track_assignment(symb = c("<-", "<+"))

untrack_assignment()

register_inline_summary(x, ...)

\method{register_inline_summary}{character}(x, ...)

\method{register_inline_summary}{numeric}(x, ...)

\method{register_inline_summary}{logical}(x, ...)

\method{register_inline_summary}{`function`}(x, ...)

\method{register_inline_summary}{environment}(x, ...)

\method{register_inline_summary}{tbl_df}(x, ...)

\method{register_inline_summary}{data.frame}(x, ...)

\method{register_inline_summary}{default}(x, ...)
}
\arguments{
\item{x}{An object of class \code{registers}}

\item{...}{Additional arguments to pass to \code{\link[=cat]{cat()}}}

\item{symb}{Symbol to use for tracked assignment, either \code{"<-"} (default) or
\code{"<+"}}
}
\value{
Invisible \code{NULL}
}
\description{
See registers

Registers for R
}
\examples{

registers()
a <- 1
set_focus("a")
registers()
show_registers()
ls(registers())

# in RStudio, go to Tools > Modify Keyboard Shortcuts...
# set registers to Alt-Shift-Cmd-R
# try Alt-Shift-Cmd-R

f <- function(x) x^2
set_focus("f")
set_focus("plot")

# in RStudio, go to Tools > Modify Keyboard Shortcuts...
# filter by "registers", change "Show registers" to Shift-Cmd-R

track_assignment()
ls()
ls(all.names = TRUE)
registers()
a <- 1
.
b <- 2
.
..
registers()
c <- 3
.
..
c(._1, ._2, ._3)
f <- function(x, y) {
  x + y
}
e <- list2env(list(a = 1, b = 2))
df <- cars
registers()



if (FALSE) { # oddities
  ..
  .. <- 10
  .

  x <- 1:3
  .
  x[1] <- 2
  x # unchanged
  . # yet 2 was evaluated and stored
}
untrack_assignment()
}
