% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_inline_summary.R, R/registers.R,
%   R/registers_package.R
\docType{package}
\name{registers_inline_summary}
\alias{registers_inline_summary}
\alias{register_inline_summary}
\alias{register_inline_summary.character}
\alias{register_inline_summary.numeric}
\alias{register_inline_summary.logical}
\alias{register_inline_summary.function}
\alias{register_inline_summary.environment}
\alias{register_inline_summary.tbl_df}
\alias{register_inline_summary.data.frame}
\alias{register_inline_summary.list}
\alias{register_inline_summary.default}
\alias{registers}
\alias{show_registers}
\alias{reset_registers}
\alias{set_registers_option}
\alias{get_registers_option}
\alias{print.registers}
\alias{package-registers}
\title{Registers inline summaries}
\usage{
register_inline_summary(x, ...)

\method{register_inline_summary}{character}(x, ...)

\method{register_inline_summary}{numeric}(x, ...)

\method{register_inline_summary}{logical}(x, ...)

\method{register_inline_summary}{`function`}(x, ...)

\method{register_inline_summary}{environment}(x, ...)

\method{register_inline_summary}{tbl_df}(x, ...)

\method{register_inline_summary}{data.frame}(x, ...)

\method{register_inline_summary}{list}(x, ...)

\method{register_inline_summary}{default}(x, ...)

registers()

show_registers()

reset_registers()

set_registers_option(...)

get_registers_option(option)

\method{print}{registers}(x, ...)
}
\arguments{
\item{x}{An object of class \code{registers}}

\item{...}{Additional arguments}

\item{option}{A named registers option to retrieve}
}
\value{
\code{x}

Invisible \code{NULL}
}
\description{
Registers inline summaries

See registers

Registers for R
}
\examples{

register_inline_summary(c("foo", "bar"))
register_inline_summary(rnorm(3))
register_inline_summary(1:3)
register_inline_summary(c(TRUE, FALSE, TRUE))

register_inline_summary(cars)
register_inline_summary(list(a = 1, b = 2))
register_inline_summary(list(1, 2))

register_inline_summary(function(x, y) x + y)
register_inline_summary(function(x, y) {
  z <- x^2
  z + y
})

e <- new.env()
assign("a", 1, envir = e)
e
ls(e)
e$a
register_inline_summary(e)
g <- new.env(parent = e)
register_inline_summary(g)
register_inline_summary(asNamespace("dplyr"))





registers()
set_register(a <- rnorm(1), "a")
registers()
ls(registers())

# in RStudio, go to Tools > Modify Keyboard Shortcuts...
# set Show registers to Alt-Shift-Cmd-R, Apply, then try it

reset_registers()
registers()

get_registers_option("clobber")
set_registers_option("clobber" = FALSE)
get_registers_option("clobber")
set_registers_option("clobber" = FALSE)
get_registers_option("clobber")
reset_registers()
get_registers_option("clobber")

}
\seealso{
\code{\link[=set_register]{set_register()}}
}
