% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registers.R, R/registers_package.R
\docType{package}
\name{registers}
\alias{registers}
\alias{addin_show_registers}
\alias{reset_registers}
\alias{print.registers}
\alias{package-registers}
\title{Registers environment}
\usage{
registers()

addin_show_registers()

reset_registers()

\method{print}{registers}(x, ...)
}
\arguments{
\item{x}{An object of class \code{registers}}

\item{...}{Additional arguments to pass to \code{\link[=cat]{cat()}}}
}
\value{
Invisible \code{NULL}
}
\description{
See registers

Registers for R
}
\examples{

registers()
a <- 1
set_focus("a")
registers()
f <- function(x) x^2
set_focus("f")
set_focus("print")

# in RStudio, go to Tools > Modify Keyboard Shortcuts...
# filter by "registers", change "Show registers" to Shift-Cmd-R


}
