% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_register.R
\name{set_register}
\alias{set_register}
\alias{set_focus}
\alias{addin_set_focus}
\alias{addin_set_focus_interactive}
\alias{glimpse_focus}
\alias{addin_glimpse_focus}
\alias{addin_glimpse_highlighted}
\title{Registers}
\usage{
set_focus(name)

addin_set_focus()

addin_set_focus_interactive()

glimpse_focus()

addin_glimpse_focus()

addin_glimpse_highlighted()
}
\arguments{
\item{name}{The binding name to focus on, as a string}
}
\value{
Invisible \code{NULL}
}
\description{
Set registers and register actions
}
\details{
In Vim-speak, \code{focus} is the unnamed register.
}
\examples{

registers()

set_focus("cars")
glimpse_focus()

# in RStudio, go to Tools > Modify Keyboard Shortcuts...
# next to Set focus, use Alt-Shift-Cmd-F
# next to Glimpse focus, use Shift-Cmd-G
# next to Set focus interactively, use Alt-Shift-F

airquality
# highlight airquality above and push Alt-Shift-Cmd-F

# press Shift-Cmd-G

# in Tools > Modify Keyboard Shortcuts...,
# set Glimpse highlighted to Alt-Shift-H, then highlight the line below and
# press Alt-Shift-H
(a <- rnorm(1))
# now evaluate this:
a




}
