% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spy.R
\name{spy}
\alias{spy}
\alias{spy.data.frame}
\alias{spy_top}
\alias{spy_top.data.frame}
\alias{spy_btm}
\alias{spy_btm.data.frame}
\alias{spy_all}
\alias{spy_all.data.frame}
\title{Spy}
\usage{
spy(...)

\method{spy}{data.frame}(x, name = deparse(substitute(x)))

spy_top(...)

\method{spy_top}{data.frame}(df, n = 20)

spy_btm(...)

\method{spy_btm}{data.frame}(x, name = deparse(substitute(x)))

spy_all(...)

\method{spy_all}{data.frame}(df)
}
\arguments{
\item{name}{The binding name to focus on, as a string}

\item{data}{A dataset (or expression that returns one)}
}
\value{
Invisible \code{NULL}
}
\description{
Spy
}
\examples{

spy(cars)
spy_top(cars)
spy_btm(cars)
spy_all(cars)
spy_all(ggplot2::diamonds)

# Shift-Cmd-H
cars

}
